Rename classes works with data from network
